Namespace user routes
