+++ app/views/hub/users/index.html.erb
@@ -19,7 +19,7 @@
       <% @users.each do |user| %>
         <tr id="user-<%= user.id %>" class="index-table__row">
           <th class="index-table__row-header">
-            <%= link_to user.name, edit_user_path(id: user) %>
+            <%= link_to user.name, edit_hub_user_path(id: user) %>
           </th>
           <td class="index-table__cell"><%= user_roles(user) %></td>
           <td class="index-table__cell"><%= user.vita_partner&.name || t("general.none") %></td>
