+++ app/views/hub/users/profile.html.erb
@@ -25,7 +25,7 @@
     </li>
 
     <li class="user-action-link">
-      <%= link_to users_path, class: "button button--small" do %>
+      <%= link_to hub_users_path, class: "button button--small" do %>
         <%= t(".users_link") %>
       <% end %>
     </li>
